+++ papiea-engine/src/main.ts
@@ -19,7 +19,7 @@ import {
 } from "./auth/authn";
 import { createOAuth2Router } from "./auth/oauth2";
 import { S2SKeyUserAuthInfoExtractor } from "./auth/s2s";
-import { Authorizer, AdminAuthorizer, PerProviderAuthorizer } from "./auth/authz";
+import { Authorizer, AdminAuthorizer, PerProviderAuthorizer, IntentWatcherAuthorizer } from "./auth/authz";
 import { ValidatorImpl } from "./validator";
 import { ProviderCasbinAuthorizerFactory } from "./auth/casbin";
 import { BadRequestError } from "./errors/bad_request_error"
@@ -67,8 +67,9 @@ async function setUpApplication(): Promise<express.Express> {
     const watchlistDb = await mongoConnection.get_watchlist_db(logger)
     const graveyardDb = await mongoConnection.get_graveyard_db(logger)
     const validator = ValidatorImpl.create()
-    const entityApiAuthorizer: Authorizer = new PerProviderAuthorizer(logger, new ProviderCasbinAuthorizerFactory(logger));
+    const entityApiAuthorizer: PerProviderAuthorizer = new PerProviderAuthorizer(logger, new ProviderCasbinAuthorizerFactory(logger));
     const adminAuthorizer: Authorizer = new AdminAuthorizer()
+    const intentWatcherAuthorizer: IntentWatcherAuthorizer = new IntentWatcherAuthorizer(logger, entityApiAuthorizer, specDb, providerDb);
     const intentfulContext = new IntentfulContext(specDb, statusDb, graveyardDb, differ, intentWatcherDB, watchlistDb, validator, entityApiAuthorizer)
     const providerApi = new Provider_API_Impl(logger, providerDb, statusDb, s2skeyDb, watchlistDb, intentfulContext, adminAuthorizer, [adminAuthorizer, entityApiAuthorizer], validator)
     const sessionKeyApi = new SessionKeyAPI(sessionKeyDb)
@@ -94,7 +95,7 @@ async function setUpApplication(): Promise<express.Express> {
     app.use(createAuthnRouter(logger, userAuthInfoExtractor));
     app.use(createOAuth2Router(logger, oauth2RedirectUri, providerDb, sessionKeyApi));
     app.use('/provider', createProviderAPIRouter(providerApi, trace));
-    app.use('/services', createEntityAPIRouter(new Entity_API_Impl(logger, statusDb, specDb, graveyardDb, providerDb, intentWatcherDB, entityApiAuthorizer, validator, intentfulContext), trace));
+    app.use('/services', createEntityAPIRouter(new Entity_API_Impl(logger, statusDb, specDb, graveyardDb, providerDb, intentWatcherDB, entityApiAuthorizer, intentWatcherAuthorizer, validator, intentfulContext), trace));
     app.use('/api-docs', createAPIDocsRouter('/api-docs', new ApiDocsGenerator(providerDb), providerDb));
     app.use(function (err: any, req: any, res: any, next: any) {
         if (res.headersSent) {