+++ papiea-sdk/typescript/__tests__/sdk_tests/provider_sdk.test.ts
@@ -6,7 +6,7 @@ import { loadYamlFromTestFactoryDir, OAuth2Server, ProviderBuilder } from "../..
 import { timeout } from "../../../../papiea-engine/src/utils/utils"
 import axios from "axios"
 import { readFileSync } from "fs";
-import { Metadata, IntentfulBehaviour, Provider, Spec, Action, Entity_Reference, Entity } from "papiea-core";
+import { Metadata, IntentfulBehaviour, Provider, Spec, Action, Entity_Reference, Entity, IntentfulStatus } from "papiea-core";
 import { Logger, LoggerFactory } from "papiea-backend-utils";
 import {kind_client, ProviderClient} from "papiea-client"
 import { Kind_Builder, ProceduralCtx_Interface, ProviderSdk, SecurityApi } from "../../src/provider_sdk/typescript_sdk";
@@ -1268,14 +1268,13 @@ describe("Provider Sdk tests", () => {
     });
 
     test("Papiea version equal to the supported version should pass", async () => {
-        expect.hasAssertions();
+        expect.assertions(1);
         const sdk = ProviderSdk.create_provider(papieaUrl, adminKey, server_config.host, server_config.port);
         try {
             const location = sdk.new_kind(location_yaml);
             sdk.version(provider_version);
             sdk.prefix("location_provider");
             location.on_create({input_schema: location_yaml}, async (ctx, input) => {
-                expect(input).toBeDefined()
                 return {
                     spec: input,
                     status: input
@@ -1302,14 +1301,13 @@ describe("Provider Sdk tests", () => {
     });
 
     test("Papiea version compatible with the supported version should pass", async () => {
-        expect.hasAssertions();
+        expect.assertions(1);
         const sdk = ProviderSdk.create_provider(papieaUrl, adminKey, server_config.host, server_config.port);
         try {
             const location = sdk.new_kind(location_yaml);
             sdk.version(provider_version);
             sdk.prefix("location_provider");
             location.on_create({input_schema: location_yaml}, async (ctx, input) => {
-                expect(input).toBeDefined()
                 return {
                     spec: input,
                     status: input
@@ -1317,6 +1315,12 @@ describe("Provider Sdk tests", () => {
             })
             await sdk.register();
             const kind_name = sdk.provider.kinds[0].name;
+            let sdk_version_list = sdk.get_sdk_version().split('.')
+            if (sdk_version_list[2] === '0') {
+                sdk_version_list[2] = '1'
+            } else {
+                sdk_version_list[2] = (parseInt(sdk_version_list[2]) - 1).toString()
+            }
             const { data: { metadata, spec } } = await axios.post(`${sdk.entity_url}/${sdk.provider.prefix}/${sdk.provider.version}/${kind_name}`, {
                     x: 10,
                     y: 11
@@ -1324,7 +1328,7 @@ describe("Provider Sdk tests", () => {
                 headers: {
                   "Content-Type": "application/json",
                   "Authorization": `Bearer ${adminKey}`,
-                  "Papiea-Version": "0.8.5"
+                  "Papiea-Version": sdk_version_list.join('.')
                 }
             });
 
@@ -1336,14 +1340,13 @@ describe("Provider Sdk tests", () => {
     });
 
     test("Papiea version incompatible with the supported version should fail", async () => {
-        expect.hasAssertions();
+        expect.assertions(1);
         const sdk = ProviderSdk.create_provider(papieaUrl, adminKey, server_config.host, server_config.port);
         try {
             const location = sdk.new_kind(location_yaml);
             sdk.version(provider_version);
             sdk.prefix("location_provider");
             location.on_create({input_schema: location_yaml}, async (ctx, input) => {
-                expect(input).toBeDefined()
                 return {
                     spec: input,
                     status: input
@@ -1351,6 +1354,12 @@ describe("Provider Sdk tests", () => {
             })
             await sdk.register();
             const kind_name = sdk.provider.kinds[0].name;
+            let sdk_version_list = sdk.get_sdk_version().split('.')
+            if (sdk_version_list[0] === '0') {
+                sdk_version_list[0] = '1'
+            } else {
+                sdk_version_list[0] = (parseInt(sdk_version_list[0]) - 1).toString()
+            }
             const { data: { metadata, spec } } = await axios.post(`${sdk.entity_url}/${sdk.provider.prefix}/${sdk.provider.version}/${kind_name}`, {
                     x: 10,
                     y: 11
@@ -1358,11 +1367,11 @@ describe("Provider Sdk tests", () => {
                 headers: {
                   "Content-Type": "application/json",
                   "Authorization": `Bearer ${adminKey}`,
-                  "Papiea-Version": "1.0.0"
+                  "Papiea-Version": sdk_version_list.join('.')
                 }
             });
         } catch (e) {
-            expect(e.response.data.error.errors[0].message).toBe("Received incompatible papiea version: 1.0.0")
+            expect(e.response.data.error.errors[0].message).toContain("Received incompatible papiea version")
         } finally {
             sdk.cleanup()
         }
@@ -1805,6 +1814,148 @@ describe("SDK + oauth provider tests", () => {
             sdk.server.close()
         }
     });
+
+    const WATCHER_TEST_SCHEMA = {
+        Location: {
+            type: 'object',
+            title: 'Location',
+            'x-papiea-entity': 'differ',
+            required: ['x', 'y'],
+            properties: {
+                x: {
+                    type: 'number'
+                },
+                y: {
+                    type: 'number'
+                }
+            }
+        }
+    }
+    test("Intent watcher check permission read should succeed", async () => {
+        expect.assertions(1);
+        const sdk = ProviderSdk.create_provider(papieaUrl, adminKey, server_config.host, server_config.port);
+        const location = sdk.new_kind(WATCHER_TEST_SCHEMA);
+        sdk.version(provider_version);
+        sdk.prefix("permissioned_intent_watcher_read_succeed");
+        location.on("x", async (ctx, entity, input) => {
+            await providerApiAdmin.post(`/${sdk.provider.prefix}/${sdk.provider.version}/update_status`, {
+                context: "some context",
+                entity_ref: {
+                    uuid: entity.metadata.uuid,
+                    kind: entity.metadata.kind
+                },
+                status: {
+                    x: 11,
+                    y: 11
+                }
+            })
+        })
+        sdk.secure_with(oauth, modelText, "xxx");
+        try {
+            await sdk.register();
+            const kind_name_local: string = sdk.provider.kinds[0].name
+            await providerApiAdmin.post(`/${ sdk.provider.prefix }/${ sdk.provider.version }/auth`, {
+                policy: `p, alice, owner, ${ kind_name_local }, *, allow`
+            });
+            const { data: { token } } = await providerApi.get(`/${provider.prefix}/${provider.version}/auth/login`);
+            const { data: { metadata, spec } } = await entityApi.post(`/${ sdk.provider.prefix }/${ sdk.provider.version }/${ kind_name_local }`, {
+                metadata: {
+                    extension: {
+                        owner: "alice",
+                        tenant_uuid: tenant_uuid
+                    }
+                },
+                spec: {
+                    x: 10,
+                    y: 11
+                }
+            },
+            { headers: { 'Authorization': `Bearer ${token}` }});
+            const { data: { watcher } } = await entityApi.put(`/${ sdk.provider.prefix }/${ sdk.provider.version }/${ kind_name_local }/${ metadata.uuid }`, {
+                spec: {
+                    x: 11,
+                    y: 11
+                },
+                metadata: {
+                    spec_version: 1
+                }
+            },
+            { headers: { 'Authorization': `Bearer ${token}` }})
+
+            const watcherApi = sdk.get_intent_watcher_client()
+            const intent_watcher = await watcherApi.get(watcher.uuid)
+            expect(intent_watcher.status).toEqual(IntentfulStatus.Active)
+        } finally {
+            sdk.server.close();
+            await providerApiAdmin.post(`/${ sdk.provider.prefix }/${ sdk.provider.version }/auth`, {
+                policy: null
+            });
+        }
+    });
+
+    test("Intent watcher check permission read should fail if owner is incorrect", async () => {
+        expect.assertions(1);
+        const sdk = ProviderSdk.create_provider(papieaUrl, adminKey, server_config.host, server_config.port);
+        const location = sdk.new_kind(WATCHER_TEST_SCHEMA);
+        sdk.version(provider_version);
+        sdk.prefix("permissioned_intent_watcher_read_fail");
+        location.on("x", async (ctx, entity, input) => {
+            await providerApiAdmin.post(`/${sdk.provider.prefix}/${sdk.provider.version}/update_status`, {
+                context: "some context",
+                entity_ref: {
+                    uuid: entity.metadata.uuid,
+                    kind: entity.metadata.kind
+                },
+                status: {
+                    x: 11,
+                    y: 11
+                }
+            })
+        })
+        sdk.secure_with(oauth, modelText, "xxx");
+        try {
+            await sdk.register();
+            const kind_name_local: string = sdk.provider.kinds[0].name
+            // Use correct policy to create entity
+            await providerApiAdmin.post(`/${sdk.provider.prefix}/${sdk.provider.version}/auth`, {
+                policy: `p, carol, owner, ${kind_name_local}, *, allow`
+            });
+            const {data: {token}} = await providerApi.get(`/${provider.prefix}/${provider.version}/auth/login`);
+            const { data: { metadata, spec } } = await entityApi.post(`/${ sdk.provider.prefix }/${ sdk.provider.version }/${ kind_name_local }`, {
+                metadata: {
+                    extension: {
+                        owner: "alice",
+                        tenant_uuid: tenant_uuid
+                    }
+                },
+                spec: {
+                    x: 10,
+                    y: 11
+                }
+            },
+            { headers: { 'Authorization': `Bearer ${token}` }});
+            const { data: { watcher } } = await entityApi.put(`/${ sdk.provider.prefix }/${ sdk.provider.version }/${ kind_name_local }/${ metadata.uuid }`, {
+                spec: {
+                    x: 11,
+                    y: 11
+                },
+                metadata: {
+                    spec_version: 1
+                }
+            },
+            { headers: { 'Authorization': `Bearer ${token}` }})
+
+            const watcherApi = sdk.get_intent_watcher_client()
+            await watcherApi.get(watcher.uuid)
+        } catch (e) {
+            expect(e.response.data.error.errors[0].message).toBe("Permission denied.")
+        } finally {
+            sdk.server.close();
+            await providerApiAdmin.post(`/${ sdk.provider.prefix }/${ sdk.provider.version }/auth`, {
+                policy: null
+            });
+        }
+    });
 });
 
 describe("SDK callback tests", () => {
@@ -1902,6 +2053,35 @@ describe("SDK callback tests", () => {
             })
         } finally {
             sdk.server.close()
+        }
+    });
+
+    test("Engine should reject incorrect entity creation format (constructor format for example)", async () => {
+        expect.hasAssertions();
+        const sdk = ProviderSdk.create_provider(papieaUrl, adminKey, server_config.host, server_config.port);
+        sdk.new_kind(location_yaml);
+        prefix = "constructor_format_provider"
+        sdk.version(provider_version);
+        sdk.prefix(prefix);
+        try {
+            await sdk.register();
+            const kind_name = sdk.provider.kinds[0].name;
+            const {
+                data: {
+                    metadata,
+                    spec
+                }
+            } = await axios.post(`${sdk.entity_url}/${sdk.provider.prefix}/${sdk.provider.version}/${kind_name}`, {
+                x: 10,
+                y: 11
+            }, {
+                headers: {
+                    "Content-Type": "application/json",
+                    "Authorization": `Bearer ${adminKey}`,
+                }
+            });
+        } catch (e) {
+            expect(e.response.data.error.errors[0].message).toEqual("Spec was not provided or was provided in an incorrect format")
         }
     });